refactor(server): split index into focused modules
index.ts kept only module state and lifecycle (construct/start/stop/
listenServers); extracted the rest into single-responsibility files:
- server-config.ts: config types, defaults and resolveServers
- server-factory.ts: protocol-based server creation
- port-binding.ts: listen with dev port fallback
- dev-registry.ts: runtime mode and endpoint registration

index.ts goes from 344 to 100 lines; no behaviour change.
